quoll/requirements.txt
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.6.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.3
Django==2.0.5
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
lxml==4.2.1
parsel==1.4.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
PyMySQL==0.8.1
pyOpenSSL==18.0.0
python-scrapyd-api==2.1.2
pytz==2018.4
queuelib==1.5.0
requests==2.18.4
Scrapy==1.5.0
scrapyd==1.2.0
service-identity==17.0.0
six==1.11.0
Twisted==18.4.0
urllib3==1.22
w3lib==1.19.0
zope.interface==4.5.0
